Move containers->comp folder, rename css->style.css, comp->index.js
