+++ src/components/App/container.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
-import App from '../components/App/App';
-import { checkLogin } from '../actions/user';
+import App from '.';
+import { checkLogin } from '../../actions/user';
 
 const mapStateToProps = state => ({
   user: state.user,
