+++ src/components/ItemEdit/style.css
@@ -1,0 +1,4 @@
+.task-list-item--done {
+  text-decoration: line-through;
+  opacity: 0.5;
+}
